Moved bin to src directory, refactored code and added success messenge
