+++ src/bin/commands/index.js
