+++ src/bin/utils/formatUtil.js
